+++ web/app/mu-plugins/base-setup/base-setup.php
@@ -1,7 +1,7 @@
 <?php
 /*
   Plugin Name:  Base Setup
-  Description:  Automatically sets site specific defaults: * Sets upload path to /media; * Sets upload path URL to example.com/media; * Disables year/month folder structure for uploads; * Sets permalink structure to post name; * Sets save/load path to ACF PRO JSON custom fileds;
+  Description:  Automatically sets site specific defaults: * Sets upload path to /media; * Sets upload path URL to example.com/media; * Disables year/month folder structure for uploads; * Sets permalink structure to post name; * Creates data api URL; * Sets save/load path to ACF PRO JSON custom fileds;
   Version:      1.0.0
   Author:       Ruslan Komjakov
   Author URI:   https://github.com/rslnk
@@ -28,66 +28,25 @@ if ('http://' . $_SERVER['SERVER_NAME'] . '/wp' == get_option('siteurl')) {
 }
 
 /*
-  JSON API endpoint
+  API data endpoint
 
+  Creates example.com/api URL and includes custom template
 
-  https://make.wordpress.org/plugins/2012/06/07/rewrite-endpoints-api
-  https://gist.github.com/joncave/2891111
+  Based on Creating an API Endpoint in WordPress by Ken Snyder
+  http://kendsnyder.com/creating-an-api-endpoint-in-wordpress/
  */
 
-add_action( 'parse_request', function() {
-    global $wp;
-    echo $wp->query_vars;
-    foreach ($wp->query_vars as $result) { echo $result->type; echo "<br>"; }
-    if ( $wp->query_vars['pagename'] == 'api/events' ) {
-        //require_once( __DIR__ . '/my/endpoint.php' );
-      echo 'test';
-      die;
-    }
-}, 0 );
+add_action( 'parse_request', 'create_api_endpoint', 0 );
 
-/*
-add_action( 'init', 'makeplugins_endpoints_add_endpoint' );
-add_action( 'template_redirect', 'makeplugins_endpoints_template_redirect' );
-
-register_deactivation_hook( __FILE__, 'makeplugins_endpoints_deactivate' );
-register_activation_hook( __FILE__, 'makeplugins_endpoints_activate' );
-*/
-
-function makeplugins_endpoints_add_endpoint() {
-  // register a "json" endpoint to be applied to posts and pages
-  add_rewrite_endpoint( 'json', EP_PERMALINK | EP_PAGES );
+function create_api_endpoint() {
+  global $wp;
+  // Create URL endpoint
+  if ( $wp->request == 'api' ) {
+    // Include custom template
+    require_once( dirname( __FILE__ )  . '/api-data.php' );
+    die;
+  }
 }
-
-function makeplugins_endpoints_template_redirect() {
-  global $wp_query;
-  // if this is not a request for json or it's not a singular object then bail
-  if ( ! isset( $wp_query->query_vars['json'] ) || ! is_singular() )
-  return;
-  // output some JSON (normally you might include a template file here)
-  // include custom template
-  include dirname( __FILE__ ) . '/json-api.php';
-  exit;
-}
-
-function makeplugins_endpoints_do_json() {
-  header( 'Content-Type: application/json' );
-  $post = get_queried_object();
-  echo json_encode( $post );
-}
-
-function makeplugins_endpoints_activate() {
-  // ensure our endpoint is added before flushing rewrite rules
-  makeplugins_endpoints_add_endpoint();
-  // flush rewrite rules - only do this on activation as anything more frequent is bad!
-  flush_rewrite_rules();
-}
-
-function makeplugins_endpoints_deactivate() {
-  // flush rules on deactivate as well so they're not left hanging around uselessly
-  flush_rewrite_rules();
-}
-
 
 /*
   ACF PRO Setup